requirements.txt
Flask-SQLAlchemy
python-magic
